Refactor semantic tests and factory resolvers
